+++ sdk/boletoHolmes/index.js
@@ -1,0 +1,7 @@
+const boletoHolmes = require('./boletoHolmes.js');
+
+exports.log = require('./log');
+exports.create = boletoHolmes.create;
+exports.query = boletoHolmes.query;
+exports.get = boletoHolmes.get;
+exports.BoletoHolmes = boletoHolmes.BoletoHolmes;